state/agent_state.py
from typing_extensions import TypedDict

# Agent state and related classes
class AgentState(TypedDict):
    question: str
    sql_query: str
    query_result: str
    query_rows: list
    attempts: int
    relevance: str
    sql_error: bool